Create FootballGame
The local County Council runs schoolchildren’s football competitions known as fanatical football.  Competitions are organized with a name and age group. Teams that want to play in the competition must register, providing information such as the age group of the team, the manager’s name and contact phone number and address. Each team has a set of players.  The player must be assigned to the team in the register that is sent to the Council.  Information required on players is their date-of-birth, name, address and name of guardian. 

The Council also keeps a register of referees.  Every week, the competition organiser arranges matches between teams in the competition.  The match has a venue (the location of the home team), a home team, an away team, a date and time and a referee.   When the game has been played, the referee records the home team’s score and the away team’s score.